+++ api/management/mutations/organization_profile.py
@@ -5,6 +5,8 @@ from kante.types import Info
 
 from karakter import models
 from api.management import types
+from api.management.authz import DENIED, assert_owner_or_admin
+from graphql import GraphQLError
 
 logger = logging.getLogger(__name__)
 
@@ -16,7 +18,13 @@ class CreateOrganizationProfileInput:
 
 
 def create_organization_profile(info: Info, input: CreateOrganizationProfileInput) -> types.ManagementOrganizationProfile:
-    organization = models.Organization.objects.get(pk=input.organization)
+    try:
+        organization = models.Organization.objects.get(pk=input.organization)
+    except models.Organization.DoesNotExist:
+        raise GraphQLError(DENIED)
+
+    assert_owner_or_admin(info, organization)
+
     profile = models.OrganizationProfile(organization=organization, name=input.name)
     profile.save()
     return profile
@@ -31,7 +39,13 @@ class UpdateOrganizationProfileInput:
 
 
 def update_organization_profile(info: Info, input: UpdateOrganizationProfileInput) -> types.ManagementOrganizationProfile:
-    profile = models.OrganizationProfile.objects.get(pk=input.id)
+    try:
+        profile = models.OrganizationProfile.objects.get(pk=input.id)
+    except models.OrganizationProfile.DoesNotExist:
+        raise GraphQLError(DENIED)
+
+    assert_owner_or_admin(info, profile.organization)
+
     if input.name:
         profile.name = input.name
     if input.avatar:
@@ -48,7 +62,11 @@ class DeleteOrganizationProfileInput:
 
 
 def delete_organization_profile(info: Info, input: DeleteOrganizationProfileInput) -> strawberry.ID:
-    profile = models.OrganizationProfile.objects.get(pk=input.id)
-    assert profile.organization.owner == info.context.request.user
+    try:
+        profile = models.OrganizationProfile.objects.get(pk=input.id)
+    except models.OrganizationProfile.DoesNotExist:
+        raise GraphQLError(DENIED)
+    if profile.organization.owner_id != info.context.request.user.id:
+        raise GraphQLError(DENIED)
     profile.delete()
     return input.id
